+++ app/src/commonMain/kotlin/it/vercruysse/lemmyapi/v0x18/datatypes/PrivateMessageId.kt
@@ -1,3 +1,3 @@
 package it.vercruysse.lemmyapi.v0x18.datatypes
 
-typealias PrivateMessageId = Int
+typealias PrivateMessageId = Long
